cloud_backend/requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
django-extensions==1.6.1
djangorestframework==3.3.2
wheel==0.24.0
